Finished debugging 1

+++ 1/1.c
@@ -17,15 +17,15 @@ double bisection(int p, int q, double (*func)(int, int, double)) {
     double b = 20;
     double xp;
     xp = (a + b) / 2;
-    int c = func(p, q , xp);
+    double c = func(p, q , xp);
     while(fabs(c) >= EPSILON)
     {
-        if((c * func(p, q , a)) > 0)
+        if((c * func(p, q, a)) > 0)
         {
             a = xp;
             b = b;
             xp = (a + b) / 2;
-            c = func(p, q , xp);
+            c = func(p, q, xp);
             //printf("%.4f\n", xp);
         }
         else
@@ -33,7 +33,7 @@ double bisection(int p, int q, double (*func)(int, int, double)) {
             a = a;
             b = xp;
             xp = (a + b) / 2;
-            c = func(p, q , xp);
+            c = func(p, q, xp);
             //printf("%.4f\n", xp);
         }
     }
